+++ CHANGELOG.md
@@ -1,8 +1,28 @@
-## [0.1.0] - 17/11/2025
+## [0.1.0] - 18/11/2025
+
+### Fixed
+- Fixed the possibility to add or modify one supply with the same name
+- Fixed the date format in supplyDetailContent screen
+- Fixed error in register workshop
+- Fixed error in view Workshops after a delete
 
 ### Added
+- US for RBAC main screens and navigation accross the app
+- Added filter and order supply batches
+
+## [0.1.0] - 17/11/2025
+
+### Changed
+ - UploadImageInput modified to not save persistantly the photo, instead to 
+   receive the URI value from the view model  
+
+## [0.1.0] - 17/11/2025
+
+- US for modify Supply with MVVM and clean architecture principles
 - Delete supply batch
 - Added filter and order products
+- Added register Beneficiary View
+- Added Consult Disabilities List
 
 ## [0.1.0] - 16/11/2025